Clean up secure_read()/secure_write() return type
The return type is ssize_t, not int, but some callers didn't handle
this properly.

The BIO callbacks are constrained by the OpenSSL API, so they take int
for the length and return int.  This is safe, since the return value
can't be greater than the input length.  To make it more clear that
this is intentional, cast the result of the
secure_read()/secure_write() call to int explicitly.

Discussion: https://www.postgresql.org/message-id/flat/f9aab072-0078-49e4-ab93-3b08086a4406@eisentraut.org

+++ src/interfaces/libpq/fe-misc.c
@@ -654,7 +654,7 @@ static int
 pqReadData_internal(PGconn *conn)
 {
 	int			someread = 0;
-	int			nread;
+	ssize_t		nread;
 
 	/* Left-justify any data in the buffer to make room */
 	if (conn->inStart < conn->inEnd)
@@ -1010,7 +1010,7 @@ pqSendSome(PGconn *conn, int len)
 	/* while there's still data to send */
 	while (len > 0)
 	{
-		int			sent;
+		ssize_t		sent;
 
 #ifndef WIN32
 		sent = pqsecure_write(conn, ptr, len);
